[check-licenses] Move location of static SDK notice file.
Cleaning up the root check-licenses directory.
golden/ used to be named as such as it held the "golden" license
patterns and texts, but those are moving soon. I believe a "static"
directory is a better name for holding static information like the SDK
notice file.

Also added a README file to this directory, explaining what each
subdirectory is for.

Bug: 105890
Change-Id: Ie42f645596a1fd5d85a7c01d76123a94224554fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714169
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>