[fxfs] Remove downcast from Allocator to SimpleAllocator.
This removes FakeAllocator in place of SimpleAllocator and a whole bunch
of dynamic polymorphism that we don't need. If we ever want to support
more than one allocator, we may have to put some of this back but it
seems unlikely.

Bug: 96077

Change-Id: I93817cb7d63b64425c3df7f8047b43bb0bdb654b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713849
Reviewed-by: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Aaron Drew <ripper@google.com>
Commit-Queue: Aaron Drew <ripper@google.com>