[appmgr] Fix namespace for components using scheme-Runners
Components launched with a Runner based on the component's URL
scheme were previously launched in an empty namespace, with access
only to any |additional_services| specified in the component
launch request.

The |additional_services| are now used to augment the Namespace in
which the component runs, and a FlatNamespace constructed which
populates the "/svc" directory from that Namespace's services.

Bug: b/241753768
Change-Id: I35d498f2496fdec036739ba36a5e9b48cd8abf0a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713965
Fuchsia-Auto-Submit: Wez <wez@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Wez <wez@google.com>