[netstack3] Add IpSock options field
Add a field to IpSock to hold options. This is currently empty, but will
be used to record multicast send options, e.g. the equivalent of
IP_MULTICAST_IF. Making this a template parameter makes it easy to limit
multicast options (and the cost of recording multicast options) to
usages that need them.

Bug: 103426, 39479
Change-Id: I2b5c188bca6f426969d0f1aa404ac47b3b98471b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712368
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>