[fxfs] Support using the unencrypted and data volumes
This feature is gated behind the "use_native_fxfs_crypto" flag which is
unset for all products at the moment.

On boot, if the "unencrypted" and "data" volumes are not set up, the
entire Fxfs partition will be reformatted and initialized with the
required volume scheme.  This implies that this change results in data
loss for any existing users of Fxfs.

Fshost supports either using the kms_stateless service for keys, or
using static (insecure) keys if that isn't available due to lack of
hardware support.   The existing zxcrypt policy knob is used to decide
where to get keys from (TEE vs null key).

Note that this CL also disables fsck-on-mount for Fxfs temporarily while
we make this work with multiple volumes (fxbug.dev/106845).

Bug: 102666
Change-Id: Icac0b2137d689925549ccc6f1bfdffdc83385427
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/712288
Reviewed-by: Drew Fisher <zarvox@google.com>
Size-Review: Saman Sami <samans@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>