[fidl][cpp] IncomingMessage -> IncomingHeaderAndMessage
This is a mechanical rename. The plan is to separate "transactional" and
"non-transactional" messages at the type level. A new EncodedMessage
will be added to just represent the body or a standalone message without
the header. Non-transactional uses of IncomingHeaderAndMessage will
switch to that.

Bug: 82681

Change-Id: I88f95e9822f5d241f2e3e68170c12f00cc4e31df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709972
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>