[debugger] Communicate target system page size.
Some operations on the client need the target system page size. This
adds that information to the "hello" message along with the architecture
and stores it in the client in the ArchInfo class.

Hooks this value up to the "stack-usage" command (removing the hardcoded
4K constant) and the "aspace" command (now additionally showing
committed bytes).

Change-Id: I41dce4a767951cd50f4f4e2bd75d693ae3c4ad74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714342
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>