[wlan][rsn] Adjust SAE retry parameters
The SAE retransmission timeout is raised from O.2 to 1 second,
and the maximum number of retries is lowered from 30 to 3. The
extended timeout will improve compatibility with APs which sometimes
take longer than expected to respond to an SAE Commit message.
The lower number of retries prevents ending up in a long cycle
of sending SAE messages to an AP that clearly will not accept them.

At this time, we don't currently have a straightforward method to test
this timeout in an integration test.

Test: Manual test with previously incompatible AP.
Fixed: b/238696147
Change-Id: I89501041090b755a0e2153837316db78bea4f64a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714102
Reviewed-by: Kiet Tran <kiettran@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>