[ffx][product] fallback to prior PB loading
Be clearer about what's failing and automatically fallback to
using the old methods of loading the PB if the local path fails.
It will report this attempt and failure to the trace log, but
doesn't tell the user. It silently goes on as if the local file
path were never attempted.

Change-Id: Icf3d3abc1c2700b122aabf6bbd1eb71bb69a6ac6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714709
Reviewed-by: Dave Roth <davidroth@google.com>
Commit-Queue: Dave Schuyler <dschuyler@google.com>