[debugger] Add unsafe stack to "stack-usage"
Accounts for the "unsafe" stack of each thread in the stack outputs.

Adds code to fetch the unsafe stack pointer for each thread.

Adds a generalized "gather" capability for coalescing multiple callbacks
into one.

Because of the additional columns, the output was reformatted somewhat.

Change-Id: I0ed0572d4eac5898df57ebe8814f926cedacfc31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714764
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>