[check-licenses] Rename filetree to directory.
The FileTree package just represents a directory / folder structure.
It was named as such because check-licenses uses it to create a tree of
file objects in memory during the traversal, but in reality it
complicates the check-licenses architecture.

This CL renames the "filetree" package to be "directory".

I haven't deleted the old filetree folder yet, as I want to keep this CL
as small as possible for easier review.

Bug: 105890
Change-Id: I94979887252ae88f267d683dfbd54b60448e2287
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714391
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>