[fit] Reorganize fields in fit::function_base
This is a follow up on CL 709609 to reorganize the private fields of
fit::function_base.

The main change is that now storage_type includes the target_ops pointer
so that "bits" field can have max_align_t alignment while the ops
field can occupy the padding left by this field if any. The previous
patch didn't actually reduce the size of the function_base due to the
introduction of the union {} with null_bits which was causing the
union itself to have the padding inside and have a size multple of
max_align_t.

This patch moves the alignas() one level up to the storage_type struct
and doesn't impose any aligment requirement on the bits field, but
since it is the first in the struct it would have the same alignment as
the rest of the struct (max_align_t). The union between bits and
null_bits now resides inside the storage_type so it can still be
constexpr initialized.

The former sizeof(storage_type) is now available as a private constexpr
value "storage_size", since the storage_type is now a bit larger, and
the few usages of it were updated.

A new private method is used to move storage_type since is a repeated
operation, in particular in swap(), which is now a bit more readable.

Change-Id: Ic372049607e089050963b693e79cc7ebf23312ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713523
Commit-Queue: Alex Deymo <deymo@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>