[benchmarks] Add metric name expectations files for iperf networking tests
This test does a large number of invocations of convertResults(), so
unfortunately we end up with a large number of expectation files.
Fortunately the files are not large.  (The test is instantiated with a
large number of parameter values, so maybe this could be pruned back
in future.)

In the test, repurpose the "label" argument to contain a fragment of
the expectation file name.

Change the expectedMetricNamesFile arg to convertResults() to be a
named argument for searchability and for consistency with other
functions.

Bug: 105202
Change-Id: Ibfd509fdd5e58187ff325bfcab4f59fd9b5d07fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714175
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>
API-Review: Gary Bressler <geb@google.com>