[ui][flatland] Allow default hit regions to have valid global positions.
Currently the default hit region is in the range (0, 0) --> (MAX, MAX).
Note that FLT_MIN is actually the smallest possible POSITIVE value, so
this is also a bug. However, the MAX float value can not be modified by
any matrix M that contains translation, orientation or scale. Any
multiplication causes rounding errors, resulting in a global hit region
that has undefined x,y, width and height.

Note that this solution is a short-term workaround, and still
suspectible to errors.

Bug: 107310

Change-Id: Ie70c95c89a4d8388834272218efa7b16b17fc438
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714516
Commit-Queue: Sahil Madan <sahilmadan@google.com>
Reviewed-by: Felipe Archondo <felipearchondo@google.com>