[lowpan-ot-driver][discovery_proxy]: Work-around for missing services.
This is a work around for what is fundamentally the same bug as
<fxbug.dev/99755>, where the child channel will unexpectedly close
when the channel that was used to create it is closed. The work-around
is to keep that parent channel alive while we are still using the
child.

Bug: 107548
Bug: 99755
Bug: b/241818894
Change-Id: I17c2870bb6d674d591e4fb9e6364ce5c775cc9ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715287
Commit-Queue: Robert Quattlebaum <rquattle@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
(cherry picked from commit b98e619afe8ff705628cee9da50007453520de68)