[kernel][x86] Save registers for suspend.
Continuing to add support for the S3 (Suspend to RAM) sleep state.

Part 2/3. This CL saves x86 registers before calling on a function to
transition to the sleep state. That function is not yet supported.

To follow is filling out that function to set the PM1 register with the
sleep bits.

Bug: 107532

Change-Id: I8eba15e1a02b8cde20678e27fdeb7d41e140134d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/587301
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Sarah Pham <smpham@google.com>