[kernel][x86] Set suspend registers.
Finish adding support for the S3 (Suspend to RAM) sleep state.

Part 3/3. This CL writes the sleep type and sleep enable bits to the PM1
control register, and the watches the PM1 event register for the wake
status to indicate the system has resumed.

Bug: 107532

Change-Id: I1f9b597575b7c192246fd8ea22b8fbc21ace490e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611881
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Sarah Pham <smpham@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>