[emu] Remove FfxConfigWrapper from instances
The FfxConfigWrapper serves as a data structure for overriding global
ffx_config values during testing. Since the introduction of this
structure, we've been introduced to the ffx_config's built-in
functionality for the same, ffx_config::query().set().

This CL continues the process of removing the FfxConfigWrapper by
switching all of our tests to the built-in. The functions being called
are not yet modified.

TESTED=fx test ffx_emulator_common_lib_test + manual verification

Change-Id: I5b913644641197a483f45c50e0d307f7b6ccb92f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715264
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>