[check-licenses] Move utility functions to separate package.
The git file handles running the external git binary to retrieve
information about a repository (URL, git commit hash, etc). This is
needed when generating the compliance worksheet, as we need to provide
hyperlinks to the upstream projects. This CL moves that functionality
into a separate package, so it can be used in multiple places across the
check-licenses project.

Same goes for the GN tool and its usage across the project.

This is mostly a copy-paste from the result/world/* directory.

Bug: 104400, 103596, 105890
Change-Id: Ic963ccf39c3f217195cb1fa32cc615c0e3592991
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714145
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>