[emu] Fix upscript parsing
The previous CL changed how the upscript path was being retrieved from
the ffx_config, and in doing so it broke the workflow if the value in
ffx_config was an empty string.

This CL handles the empty string case as well as the None case, and adds
a test to guard future changes.

Bug: 105286

Change-Id: I8acb4a2af4a50dcfd346bd1997d5bcf84593bd96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715514
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>