[ffx][self-test] Correct daemon-config-flag test
This makes it so that the following happen in the daemon config-flag
test that were previously not happening, and likely leading to
false-positives of this test (as well as sometimes false negatives):

- Wait to make sure the daemon actually spawns and stays spawned.
- Wait to make sure the daemon doesn't die after the quit message is
  sent to the 'wrong' daemon.
- Don't do a blocking wait in the async block, preventing things from
  moving forward (Child::wait).

This change also makes it so that the output for self-test changes to
show that a test has started and then only prints that status after it's
done. This is just useful to make it so that you know the test really
started and that it's not stuck on cleanup on the previous test or
something.

Multiply: ffx-e2e-without-target
Fixed: 105903
Change-Id: Ie08baf8f8671fcb3b568198aa53c10b3b505e01c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715605
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Andrew Davies <awdavies@google.com>