[emu] Remove FfxConfigWrapper from engine tests
The FfxConfigWrapper serves as a data structure for overriding global
ffx_config values during testing. Since the introduction of this
structure, we've been introduced to the ffx_config's built-in
functionality for the same, ffx_config::query().set().

This CL continues the process of removing the FfxConfigWrapper by
switching all of our tests to the built-in. The functions in the engine
module have had FfxConfigWrapper removed, except for the engine.run
function. The calls to get_host_tool have been wrapped and mocked, since
there's no easy way to override those values in the ffx_config.

This should be the last of the tests that need updated before we can
remove the data structure entirely.

TESTED=fx test ffx_emulator_engines_lib_test + manual verification

Bug: 105286

Change-Id: I2bb06b082a08353e05789a446298345eb429485d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715520
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>