[kernel][lib/arch] Use retry logic in arch::Random
The hw rng instructions (`rdrand`/`rdseed` on x86-64 and `rndr`/`rndrrs`
on arm) might fail if the system is under heavy load.

This commit adds a retry logic to these instructions, allowing users to
specify the amount of retries, and providing a default value that should
be good enough for regular use.

[0]: Intel Digital Random Number Generator (DRNG) Software Implementation Guide

Bug: 106916
Change-Id: I49ed8fd876dbbdce855729bc1343c0b9ad27dd2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715243
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>