[cobalt] Enable the test_dont_backfill_empty_reports flag in the testapp.
The testapp needs the backfilling of reports that have never had any events
to be disabled, as it relies on that bug to function correctly. With
backfilling enabled, the first test case the testapp runs causes a backfill
of all reports in the registry. This includes setting the last sent time
such that future observation generation will not run again for those
reports. But since the testapp doesn't clear the local aggregate data
between test cases, the next test case that runs will result in no
observations being generated. Also, the initial backfill of the entire
registry would generate a few thousand observations for all the private
reports in the registry, filling the observation store and causing the
test to fail.

Bug: 107221
Change-Id: I24eadc4d6db9baed67ab466f470d70ada26422c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715224
Commit-Queue: Cameron Dale <camrdale@google.com>
Reviewed-by: Zach Bush <zmbush@google.com>