[hypervisor][x86] Change how we perform invalidations
Firstly, make sure that we invalidate an EPT across all CPUs when we
first create it, to ensure that if we are reusing the address of a
previous EPT that all cached data is invalidated.

Secondly, don't invalidate EPT (host physical to guest physical)
translations when we migrate CPUs or when we create the VCPU, only
invalidate VPID translations.  The EPT only needs to be invalidated when
we modify it, but we may have stale VPID translations cached on the
target CPU.

Multiply: hypervisor_tests
Multiply: virtualization_core_tests
Bug: 95763
Change-Id: I8ae1ac9e51614c0875bc78673af89efcc0b7cccc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715462
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>