[line_input] Update reverse-search prompt.
Previously line_input copied bash's Control-R prompt which was:
  (reverse-i-search)`text': text that matches
The use of the backtick was presumably to emulate smart quotes. But most
modern terminal fonts have a very different presentation for the
backtick character that looks odd.

This change uses ' for both, so the prompt is now:
  (reverse-i-search)'text': text that matches

Change-Id: I98b4cacb624041c4fb9dd650c94903f261859ce6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714382
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>