[debugger] Optionally pretty-print lambdas.
Puts the existing lambda pretty-printing for identifiers behind an
option so it can be disabled. Disable this from the "verbose" thread
backtrace mode so the symbols can be matched to other sources.

Use "operator" syntax highlighting for name separators ("::") when
printing function names and identifiers. The existence of this syntax
highlighting class is newer than the name formatting code.

Change-Id: Ia2bd96cbcfbf05b979a318f1b3ae6d5b40f015a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714782
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>