[cobalt] Filter test projects out of the production registry.
A separate set of config data containing the full registry (with test
projects) is provided to the testapp.

Allowlist the cobalt_tests package to receive config_data. This is
not a new use of config_data, we are just providing the same set of
config_data for the Cobalt component, but using different data for
the test.

Bug: 107294
Change-Id: Ic3d143ac89232413de5f24e989af799645c2d3ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715131
Reviewed-by: Zach Bush <zmbush@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Cameron Dale <camrdale@google.com>