[starnix] Ensures HandleTable::insert_for_transaction ref integrity
HandleTable::insert_for_transaction can either returns a new handle with
a strong ref count of 1, or return an existring handle. In the second
case, before this CL, it didn't change the ref count of the returned
handle. As the caller was expecting to own a strong reference, it would
decrease it, and delete handles that were still referenced.

This CL updates the behaviour by ensuring that the strong ref count of a
returned handle is increased.

Change-Id: If7b7ef8ec671aaf79c49e89cbfba8d3a0b790ac7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715866
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>