[tpm] Add ExecuteCommand to the TPM Driver.
The existing ExecuteVendorCommand constructs the TPM header and
assumes a header will not be provided. This works well for non-session
commands but session based commands require tags to be written into
the header section of the TPM. This change introduces a new method
that allows the driver to accept a full command with a header. This
change also deprecates ExecuteVendorCommand and we should aim
to remove it in a future revision once the call sites have been
updated.

Bug: 105028
Change-Id: I19b604f7e8905c25769786f4c6857202f33b8ebe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709811
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Ben Wright <benwright@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>
API-Review: Jody Sankey <jsankey@google.com>