[bt][fastpair] Use big-endian address in GATT notifications
https://developers.google.com/nearby/fast-pair/specifications/introduction#OctetOrder
notes that all multi-byte fields should be big endian.

Fix a bug where we are not sending the local address in big endian.

While here, add a warning if the received local address differs from
that of the local device. Make this failure non-fatal and mark with a
TODO for a future update.

Bug: 107656
Test: fx test bt-fastpair-provider-tests
Test: Manual - verify via Wireshark that the outgoing GATT notification
has the address correctly formatted in Big-E.

Change-Id: Id0facf9f3ad5780154caf4ef905d15f5a99f4c2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715727
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>