[ffx][config] Context-awareness in ffx config (v3)
Re-landing this after major changes to lockfile handling to make it more
reliable and better able to deal with "unusual" circumstances.

This also removes (compared to earlier attempts to merge) the long-hold on
the environment lockfile that was intended to prevent overlapping
read/modify/write conflicts. That seemed to cause a bunch of unrelated
problems in lockfile handling that I believe have been largely fixed now,
but I want to land that in a separate change later anyways to isolate
its effect.

This is a reland of commit 21860e901fcd68fd37fea69cb10852adbcdf846a
Which was a reland of commit 4c113119d7a105b575544010aea20431abc46597
>
> Original change's description:
> > [ffx][config] Context-awareness in ffx config
> >
> > This makes it so that ffx config is aware of the context in which it's
> > being run, understanding to start three main contexts it can exist in:
> >
> > - In a fuchsia.git build tree
> > - Isolated for testing purposes
> > - Outside a build tree just being run on its own.
> >
> > This change doesn't really change any behaviour based on this
> > understanding, but it does do some extra cleanup of how environments are
> > constructed and managed.
> >

Multiply: ffx-e2e-without-target
Bug: 106578
Bug: 82518
Change-Id: I6a03fe6f6287a20e40e3b7bc69765d8c89147125
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711545
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>