[netstack3] Put sockets in accept queue only when it's SYN_RCVD
Previously we put a socket in `LISTEN` state into the SYN queue if the
incoming segment has and only has a SYN flag, it is guaranteed to
transition into SYN_RCVD but the logic is subtle. This change refactors
this and made it explicit that only SYN_RCVD sockets can go into the
queue. This change also makes it easier to install a retransmission
timer for the SYN-ACK we sent, which will be done in the next commit.

Fixed: 107592

Change-Id: I99c9231eafb08bbacc54113f400c7ab42ccee773
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715279
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>