[golden_go_proto] Ignore comment whitespace
The diff is sensitive to whitespace changes in comments, which makes
it more difficult to roll Go versions due to changes in the formatting
of generated Go code for protos. This CL changes the script to remove
any whitespace in comments before comparing.

Also updates the error message such that the suggested `diff`
invocation no longer ignores whitespace, since this was causing `diff`
to return nothing in cases where the golden verification script was
finding differences. (Now, `diff` is at least as sensitive as the
verification script, which is what we want.)

Bug: 93043
Test: verify_golden_go_proto_file_test
Change-Id: I3fac55b79b9a4f4717da2332b27433439ff9c148
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716153
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>