[component_manager] Append `OpenFlags::Directory` when opening directory capabilities
There is a bug with routing subdirectories of `/dev` to components.
Components would not be able to open these capabilities, experiencing
PEER_CLOSED errors or other vague errors like `Broken Pipe`.

When opening subdirectories of `/dev`, devfs cares about whether
the `OpenFlags::Directory` flag is passed in. This flag determines
whether the subdirectory should be opened as a directory or service.

Component Manager did not pass this flag, so devfs would attempt
to open the subdirectories as a service, which means they would
not respond to `fuchsia.io` calls (if they opened at all).

It is worth noting that Component Manager does pass in
`MODE_TYPE_DIRECTORY`, but this looks to be ignored by devfs.

This CL ensures that the `OpenFlags::Directory` flag is now passed
in when opening any directory/storage capability, adding stronger
guarantees about what type of capability Component Manager is asking
for from the component.

Bug: 107862, 106028
Run-All-Tests: true
Change-Id: I269df60b1df0e3cbf6d16cc1f174d912ece35b8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716503
Reviewed-by: Luke Nicholson <lukenicholson@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>