[build] Allow goldens to be reformatted before diffing
It can be useful to have goldens less sensitive to formatting
differences (e.g., to prevent breakage in the case where a new version
of the formatter is being rolled into the project). This was also
a feature of `golden_test()` desired for `golden_files()`.

As a proof-of-concept, we migrate the libfuzzer golden diffing over to
using `golden_files()` with the formatting specification.

Change-Id: I49580ce0b147c0a2aaf0ffc13b781206760a5947
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715738
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>