[performance] Remove Insntrace Binary
Insntrace isn't used as far as I'm aware, and the actual implementation
is experimental and Intel only. In addition, between the garnet
migration and cfv2 migration, and now a ktrace migration, it's been
using up time to keep it alive for no benefit.

This removes the insntrace component that interfaces with the driver,
which is up next to be removed.

Bug: 107507
Change-Id: Ibd998d912e0e5c06ee5cb1de1f83c77c3e6d6e2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715512
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>