[kernel][vm] Option to place anonymous reclaim queue
Adds an option to place anonymous pages in the reclaim queue and
support for aging them. With this option being a compile time
constant, none of the added code presently exists in final builds
and is effectively a no-op.

This option is currently harded coded to false due to both lack of
tests, and the ASSERT syscalls/object.cc that would trip if enabled.

Bug: 60238
Change-Id: I1cf7bbbeba061e035048bb0afeb4a4cf3ae9f5a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715786
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>