[component_manager] Bind routing tasks to component instance
Instead of binding routing tasks to the namespace's lifetime, bind them
to the component instance instead.

This reduces the probably that a message sent to a component's namespace
at the end of its execution gets dropped because the deliver of the
request raced with component teardown.

This is a fix sufficient to address flakes that emerged with the
migration to event_streams (fxbug.dev/107764). However, it is not
bulletproof. If the component is destroyed it will still cancel the
task, and it's still possible (albeit much less likely) for a routing
request to be dropped if the namespace VFS hasn't yet processed the
request from the channel.

Followup work should be done to investigate a more complete and robust
solution. For example, we may want to wait for all tasks to terminate,
possibly in combination with a timeout.

Bug: 76579
Bug: 81980
Bug: 107764

Change-Id: I7c6e5a6494438cdb85bac14146d7357203a98e80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716168
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>