[zither] Fix _golden_test FIDL for new go/format
zither_golden_test is sensitive to formatting changes in generated Go
source, which is blocking rolling Go to 1.19 due to changes in
go/format. As a workaround, this adjusts the FIDL source such that the
generated Go is formatted the same way before and after 1.19.

Additionally, this change adds support for ignoring whitespace changes
to golden_files, as the new go/format has different behavior not only
with respect to blank comment lines but also with respect to tabs vs.
spaces in comment indentation.

Bug: 93043
Bug: 107996
Change-Id: I2eaf423e3a1a40dd0a008f93f9159fc4b774c2d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717049
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>