[syscall][pci] Avoid a potential overflow.
When verifying the size of the PCI window descriptions in
sys_pci_init, be careful how we perform the calculation of how large
the flexible portion of the structure is.  This calculation depends on
a value passed by the user in the structure itself which cannot be
fully trusted.

So:
++ Change the multiplication used here to one which tests for overflow
   instead.
++ Change the addition which could hypothetically overflow into a
   subtraction operation which cannot underflow (because of a previous
   check).

Change-Id: I30f00bd8e36d49a73ded3d495c2d344e85f2b8a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716818
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Chris Palmer <palmer@google.com>