[performance] Remove insntrace from cpu-trace driver
Insntrace isn't used, and the actual implementation is experimental and
Intel only. This removes the the ability for the cpu-trace driver to
connect to the insntrace fidl protocol making it solely a pmu driver.

The fidl protocol itself will be removed in a following patch.

Change-Id: I44e2a47f0427332ed766c2aedf7d177659fbe4a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715521
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>