[debugger] Job cannot be attached or detached
Since Zircon now supports multiple debug exception channels on one job,
it's safe to assume that the debug_agent can always create the debug
exception channel on the root job, which eliminates the need to manual
attach or detach a job.

Remove the noun "job" and the support for jobs in "attach" and "detach".
Remove the Job class in zxdb and DebugJob class in debug_agent.

Change-Id: Ic790a979f3c9ce08bb453f44a4e17b08f0e06307
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715737
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>